+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added `elIndex` and `i` to `factory` task attribute expressions.
 - expressions support object variables.
 - expressions support variable array one-dimension indices.
+- `chart` support default values.
 
 ## [0.2.13] -
 - A version bump was required due using npm publish with npm@5.6.0. See [0.2.10].